store helper function in @utils
